Album covers work

+++ Album.py
@@ -1,5 +1,6 @@
 import os
-from tkinter import Image, PhotoImage
+from tkinter import PhotoImage
+from PIL import Image
 class Album:
     def __init__(self, path=None):
         jpath = os.path.join(path, 'cover.jpg')
@@ -10,7 +11,7 @@ class Album:
             cover = Image.open(jpath)
             w, h = cover.size
             if w > 140:
-                ratio = (140, h * (140 / w))
+                ratio = (140, int(h * (140 / w)))
                 cover = cover.resize(ratio)
                 cover.save(ppath)
         if os.path.exists(ppath):
